Refactored the tests. All of them were disabled and changed position in the file structure.
