+++ TeamCode/src/main/java/org/nknsd/teamcode/programs/tests/allYears/autonomousTests/MoveToPosTest.java
@@ -1,4 +1,4 @@
-package org.nknsd.teamcode.programs.tests;
+package org.nknsd.teamcode.programs.tests.allYears.autonomousTests;
 
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
