+++ TeamCode/src/main/java/org/nknsd/teamcode/programs/tests/allYears/physicalComponent/MotorSpeedTester.java
@@ -1,10 +1,9 @@
-package org.nknsd.teamcode.programs.tests;
+package org.nknsd.teamcode.programs.tests.allYears.physicalComponent;
 
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.hardware.HardwareMap;
 
 @TeleOp(name = "Speed Test") @Disabled
 public class MotorSpeedTester extends LinearOpMode {
