+++ TeamCode/src/main/java/org/nknsd/teamcode/programs/tests/allYears/sensor/FlowSensorTestProgram.java
@@ -1,5 +1,6 @@
-package org.nknsd.teamcode.programs.tests;
+package org.nknsd.teamcode.programs.tests.allYears.sensor;
 
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
 import org.nknsd.teamcode.components.handlers.odometry.AbsolutePosition;
@@ -9,7 +10,7 @@ import org.nknsd.teamcode.frameworks.NKNProgram;
 
 import java.util.List;
 
-@TeleOp(name = "Flow Sensor Test", group="Tests")
+@TeleOp(name = "Flow Sensor Test", group="Tests") @Disabled
 public class FlowSensorTestProgram extends NKNProgram {
     @Override
     public void createComponents(List<NKNComponent> components, List<NKNComponent> telemetryEnabled) {
