+++ TeamCode/src/main/java/org/nknsd/teamcode/programs/tests/allYears/stateMachine/StateMachineTester.java
@@ -1,4 +1,4 @@
-package org.nknsd.teamcode.programs.tests;
+package org.nknsd.teamcode.programs.tests.allYears.stateMachine;
 
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
