+++ TeamCode/src/main/java/org/nknsd/teamcode/programs/tests/thisYear/firing/LauncherTest.java
@@ -1,4 +1,4 @@
-package org.nknsd.teamcode.programs.tests;
+package org.nknsd.teamcode.programs.tests.thisYear.firing;
 
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
