+++ TeamCode/src/main/java/org/nknsd/teamcode/programs/tests/thisYear/lift/ManualLiftTest.java
@@ -1,19 +1,19 @@
-package org.nknsd.teamcode.programs.tests;
+package org.nknsd.teamcode.programs.tests.thisYear.lift;
 
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.CRServo;
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
-import org.nknsd.teamcode.components.handlers.LiftHandler;
 import org.nknsd.teamcode.components.utility.StateMachine;
 import org.nknsd.teamcode.frameworks.NKNComponent;
 import org.nknsd.teamcode.frameworks.NKNProgram;
 
 import java.util.List;
 
-@TeleOp(name = "manual lift test", group = "Tests")
+@TeleOp(name = "manual lift test", group = "Tests") @Disabled
 public class ManualLiftTest extends NKNProgram {
 
     public class LiftTheServos extends StateMachine.State{
