+++ TeamCode/src/main/java/org/nknsd/teamcode/programs/tests/thisYear/microwave/MicrowaveScoopTester.java
@@ -1,5 +1,6 @@
-package org.nknsd.teamcode.programs.tests;
+package org.nknsd.teamcode.programs.tests.thisYear.microwave;
 
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
@@ -14,7 +15,7 @@ import org.nknsd.teamcode.frameworks.NKNProgram;
 
 import java.util.List;
 
-@TeleOp(name = "Microwave Scoop Tester", group = "Tests")
+@TeleOp(name = "Microwave Scoop Tester", group = "Tests") @Disabled
 public class MicrowaveScoopTester extends NKNProgram {
 
     class SlotSwitchState extends StateMachine.State {
