+++ TeamCode/src/main/java/org/nknsd/teamcode/programs/tests/thisYear/systemsTest/AutomaticPositioningTest.java
@@ -1,5 +1,6 @@
-package org.nknsd.teamcode.programs.tests;
+package org.nknsd.teamcode.programs.tests.thisYear.systemsTest;
 
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
@@ -29,7 +30,7 @@ import org.nknsd.teamcode.frameworks.NKNProgram;
 
 import java.util.List;
 
-@TeleOp(name = "automatic positioning tester", group = "Tests")
+@TeleOp(name = "automatic positioning tester", group = "Tests") @Disabled
 public class AutomaticPositioningTest extends NKNProgram {
 
     TargetingSystem targetingSystem = new TargetingSystem(RobotVersion.INSTANCE.aprilTargetingPid);
